Update README.md
This was my first attempt to create a webpage for myself based on the learnings from HTML and CSS. I used it for a year before deciding to take it down as it was static plus not so appealing. I will be creating a new one using a framework (Bootstrap) to make it more aesthetic and appealing to the eye.